discord_bot/requirements.txt
discord.py==1.7.3
asyncio==3.4.3